# docs/swagger/swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/EqualNine/Lab01/1.0.0
info:
  version: "1.0.0"
  title: ServerING API
  description: The API for the ServerING project
paths:
...
  '/api/v1/players':
    get:
      tags:
          - Player
      description: Returns All Players

      responses:
        '200':
          description: All Players

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
                
        '204':
            description: No Content
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No Players"
    
    post:
      security:
        - bearerAuth: []      

      tags:
        - Player
      description: Post Player

      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Authorization Bearer

      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  example: Player
                lastPlayed:
                  type: string
                  example: 2022-10-10
                hoursPlayed:
                  type: integer
                  example: 23

      responses:
        '201':
          description: Saved Player

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
                
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "User must be logged in"
        
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "Not enough rights for this"
  '/api/v1/players/{id}':
    get:
      tags:
          - Player
      description: Return player by Id

      parameters:
          - in: path
            required: true
            name: id
            schema:
              type: integer
            description: Player Id

      responses:
        '200':
          description: Player by Id

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        
        '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No player with id: '5'"

    put:
        security:
          - bearerAuth: []
          
        tags:
          - Player

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
        
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: Player Id
        
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  nickname:
                    type: string
                    example: Player
                  lastPlayed:
                    type: string
                    example: 2022-10-10
                  hoursPlayed:
                    type: integer
                    example: 22

        responses:
          '201':
            description: Updated Player

            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/Player'
          
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No player with id: '5'"

    patch:
        security:
          - bearerAuth: []
          
        tags:
          - Player

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
        
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: Player Id
        
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: Player

        responses:
          '201':
            description: Updated Player 
            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/Player'

          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No player with id: '5'"

    delete:
        security:
          - bearerAuth: []

        tags:
          - Player
        description: Delete Player by Id

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
    
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: Player Id
                
        responses:
          '200':
            description: Player removed
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: "player deleted successfully!"
          
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No player with id: '5'"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # только для документации
  schemas:
...
    Player:
      type: object
      properties:
        nickname:
          type: string
          example: Player
        lastPlayed:
          type: string
          example: 2022-10-10
        hoursPlayed:
          type: integer
          example: 23